third_party/blink/web_tests/external/wpt/css/css-tables/fixed-layout-2-expected.txt

This is a testharness.js-based test.
[PASS] Table-layout:fixed is not applied when width is auto
[PASS] Table-layout:fixed reports fixed when width is auto
[PASS] Table-layout:fixed is not applied when width is max-content
[PASS] Table-layout:fixed reports fixed when width is max-content
[FAIL] Table-layout:fixed is applied when width is min-content
  assert_equals: expected 100 but got 50
Harness: the test ran to completion.
